microservices/base/5-videoasset/requirements.txt
boto3
flask
python-dotenv
requests
pytube